local dataset creation essentially working.
